+++ esp_cam_sensor/CHANGELOG.md
@@ -1,3 +1,8 @@
+# Changelog
+
+## Unreleased
+- Changed YUV422 sequence for SC030IOT, SC101IOT, GC0308, and BF3A03 sensors to depend on CONFIG_CAMERA_SENSOR_SWAP_PIXEL_BYTE_ORDER
+
 ## 1.2.1
 - Deprecated ledc intr type config